finish animation stuff
