+++ src/lib/images/RoadmapFooterMobile.svelte
@@ -1,3 +1,9 @@
+<script lang="ts">
+    import { createSpringAnimation } from '$lib/utility/springAnimation';
+
+    const { springPosition, handleMouseEnter, handleMouseLeave } = createSpringAnimation();
+</script>
+
 <svg
     width="722"
     height="267"
@@ -6,7 +12,14 @@
     xmlns="http://www.w3.org/2000/svg"
     class="animate-flying-object"
 >
-    <g filter="url(#filter0_d_1043_14696)" class="animate-bear-object-mobile">
+    <!-- svelte-ignore a11y_no_static_element_interactions -->
+    <g
+        filter="url(#filter0_d_1043_14696)"
+        class="animate-bear-object-mobile"
+        onmouseenter={handleMouseEnter}
+        onmouseleave={handleMouseLeave}
+        style="transform: translateY({springPosition.current.y}px);"
+    >
         <path
             d="M457.74 238.02C455.635 231.892 455.754 195.923 460.235 182.372C460.553 181.41 460.359 180.346 459.661 179.612L457.348 177.177C453.712 173.35 451.002 168.741 449.425 163.704L448.461 160.627C447.543 157.695 448.084 154.501 449.917 152.035L455.893 143.991C456.358 143.365 456.457 142.539 456.153 141.82L455.563 140.427C453.927 136.564 454.485 132.12 457.024 128.78L457.336 128.37C458.597 126.711 460.342 125.483 462.329 124.856C465.248 123.936 468.427 124.39 470.97 126.092L473.468 127.764C474.004 128.123 474.713 127.685 474.634 127.046L474.547 126.349C474.329 124.602 475.292 122.921 476.91 122.226C478.085 121.72 479.426 121.835 480.612 122.313C483.827 123.607 487.989 124.206 487.989 124.06V123.502C487.989 121.232 489.779 119.365 492.047 119.27C496.261 119.094 500.432 120.171 504.035 122.364L511.048 126.633C511.553 126.941 512.193 126.918 512.674 126.574L513.957 125.661C516.885 123.575 520.691 123.176 523.988 124.608L525.717 125.359C529.942 127.195 531.886 132.103 530.063 136.334C529.964 136.564 530.133 136.82 530.383 136.82H544.54L552.002 135.864C560.736 134.745 569.568 134.591 578.336 135.406C585.133 136.038 591.533 138.889 596.549 143.52L598.196 145.041C599.069 145.848 600.446 145.785 601.493 145.223C603.619 144.081 607.372 143.49 610.735 150.24C614.163 157.122 611.104 160.85 608.796 162.531C607.735 163.304 607.035 164.628 607.367 165.898C610.192 176.717 611.129 178.336 613.804 194.24C616.112 207.961 616.119 223.606 615.922 235.565C615.842 240.412 613.374 244.907 609.346 247.603C607.1 249.106 604.486 249.965 601.787 250.088L587.173 250.755C583.742 250.912 580.395 249.662 577.906 247.295C577.734 247.132 577.504 247.046 577.267 247.057L562.968 247.702C559.892 247.84 556.905 246.649 554.77 244.431C552.847 242.435 551.773 239.772 551.773 237V222.254C551.773 220.702 550.404 219.507 548.866 219.718L546.557 220.034C540.403 220.876 534.184 221.153 527.98 220.861L525.598 220.749C524.282 220.687 523.133 221.634 522.981 222.943C522.438 227.606 521.883 235.115 522.621 240C523.717 247.26 512.538 249.24 504.428 249.9C498.563 250.377 494.302 248.668 491.564 246.77C490.193 245.82 488.046 245.954 486.473 246.51C484.344 247.263 480.581 247.742 473.303 248.14C461.247 248.8 459.932 244.4 457.74 238.02Z"
             fill="white"
@@ -277,7 +290,7 @@
         />
     </g>
     <path
-        opacity={1}
+        opacity={0}
         d="M22.4464 43.8801C24.861 47.5021 24.6935 53.7895 24.8341 57.8954C25.2384 69.7002 23.4767 81.451 21.0344 92.9647C19.2849 101.213 17.5694 109.524 9.38622 113.375C7.6139 114.209 4.62071 114.501 3.19875 115.638C2.4075 116.271 1.15318 119.503 2.4305 120.123C4.28114 121.022 6.80386 120.312 8.68027 119.915C12.2236 119.166 15.7314 118.134 18.3352 115.472C22.7343 110.974 24.3386 103.662 26.3706 97.9271C28.8077 91.0492 31.944 84.2776 33.6378 77.1638C35.1294 70.899 36.5481 62.1469 34.406 55.8398C33.5692 53.3759 32.6078 51.786 31.25 49.6523C30.2459 48.0745 29.3662 47.5202 27.8863 46.3095C26.1648 44.9009 26.084 43.5144 25.1456 41.6377C24.6496 40.6458 23.5055 41.6056 22.9031 40.1428C22.784 39.8535 23.5777 39.5764 23.7544 39.1876C24.3352 37.9099 23.9413 35.8613 23.9413 34.5159C23.9413 15.3726 25.1456 21.2037 25.1456 1.98584"
         fill="none"
         stroke="#201617"
@@ -287,7 +300,7 @@
         class="animate-balloon-path"
     />
     <path
-        opacity={1}
+        opacity={0}
         d="M40.3469 2.56723C32.8588 2.56723 25.0647 4.13276 19.1494 8.87709C10.6969 15.6565 6.08424 24.5992 3.43141 34.7306C-0.526321 49.8455 3.48243 64.8368 12.9277 77.1846C18.4069 84.3475 28.4246 88.7216 37.0068 91.1383C39.3904 91.8095 42.0403 91.7523 44.4947 91.7523C46.9929 91.7523 49.8953 92.1271 52.3319 91.5194C55.0881 90.832 57.5366 89.4474 60.1909 88.5339C62.4984 87.7397 64.1613 86.1132 66.2598 84.8919C69.4615 83.0287 71.579 79.4619 73.573 76.507C76.5903 72.0359 78.3198 66.7724 79.6419 61.6428C82.3532 51.1238 83.2564 39.0878 77.4589 29.4159C71.9191 20.174 66.4147 11.0202 56.5452 5.61629C51.408 2.80349 46.2142 1.42383 40.3469 1.42383"
         fill="none"
         stroke="#201617"
@@ -297,7 +310,7 @@
         class="animate-balloon-path-2"
     />
     <path
-        opacity={1}
+        opacity={0}
         d="M5.48167 1.79639C2.87329 1.94982 -1.46075 9.04611 2.88327 8.25629C4.33886 7.99164 5.51855 6.89872 6.61846 6.01878C7.20958 5.54589 5.51118 3.61033 5.48167 2.84296C5.43448 1.61628 5.98563 1.68066 6.85304 2.51816C7.45972 3.10392 10.6424 5.28731 10.6424 6.16314C10.6424 6.88737 7.16475 5.42843 6.97935 5.38723C5.79053 5.12305 4.50727 5.17319 4.50727 3.74519"
         stroke="#201617"
         stroke-width="2"
@@ -306,7 +319,7 @@
         class="animate-balloon-path-3"
     />
     <path
-        opacity={1}
+        opacity={0}
         d="M14.7839 11.2862C13.5538 12.7023 12.4035 14.189 11.0176 15.4653C10.1044 16.3064 8.76302 17.0881 8.15037 18.2095C7.91188 18.646 7.92618 19.0792 8.4784 18.9647C10.6933 18.5058 12.9109 15.5701 14.5652 14.131C15.747 13.1029 16.8484 11.9788 18.0156 10.9337C18.7938 10.2369 19.7258 9.81987 20.5305 9.17143C22.6079 7.49735 16.6263 12.8662 15.0026 15.0122C13.395 17.1367 11.9446 19.3877 10.3615 21.5327C9.06727 23.2863 7.55601 24.9246 6.40087 26.7818C6.342 26.8764 5.31243 28.6987 5.87845 28.3175C7.5838 27.169 8.97947 24.9734 10.24 23.3705C15.1683 17.1038 20.4564 11.8638 26.8724 7.24549C28.4965 6.07646 30.1528 4.97534 31.8779 3.97266C32.8097 3.43106 32.479 3.51003 31.8172 3.96007C30.4595 4.88322 29.395 6.25963 28.306 7.48466C25.354 10.8054 22.575 14.3263 19.8623 17.857C16.9208 21.6854 14.1009 25.6107 11.0662 29.3623C8.60745 32.4018 5.89665 35.1872 3.38785 38.1737C3.34348 38.2266 1.86464 40.0044 2.50095 39.5458C4.58747 38.0419 6.44203 36.1079 8.27186 34.2967C16.2199 26.4294 23.3912 17.8767 30.6508 9.33507C32.4969 7.163 34.1582 4.88758 35.8143 2.56283C36.7525 1.24579 36.1451 2.38685 35.6806 3.05375C31.1703 9.52935 26.8426 16.1394 22.3286 22.6152C19.3001 26.9599 16.1862 31.2087 12.8764 35.3289C9.79371 39.1665 6.4412 42.8485 3.63084 46.9097C3.5739 46.992 1.76747 49.8033 1.90564 49.8804C2.65766 50.3 5.08001 47.2372 5.42893 46.8593C12.757 38.9238 19.4956 30.4726 26.4715 22.2124C29.3118 18.8491 32.4704 15.8127 35.3526 12.4946C36.0799 11.6573 36.7505 10.7793 37.4301 9.90152C37.6805 9.57811 38.5319 8.84307 38.1348 8.89449C37.5956 8.96432 36.5992 11.0014 36.3367 11.3869C34.0305 14.7741 32.0776 18.3934 29.849 21.8348C24.3625 30.3069 18.3165 38.2534 12.354 46.3558C10.5292 48.8356 8.77067 51.3803 7.00834 53.9085C6.31585 54.9019 5.11753 56.2914 6.76535 54.7393C13.6404 48.2636 19.9275 41.3027 25.7911 33.8184C29.4838 29.1051 33.1512 24.4613 37.2843 20.148C39.5901 17.7418 42.1136 15.5295 44.3431 13.0485C44.9687 12.3522 44.5462 12.401 44.0515 12.9478C42.2957 14.8882 41.0488 17.3846 39.5684 19.5438C32.9444 29.2046 25.9111 38.4652 18.7688 47.7153C18.3347 48.2776 8.62027 61.3223 10.07 61.9899C10.6276 62.2467 12.2521 59.6756 12.4755 59.3968C15.6047 55.4909 18.9521 51.7831 22.0856 47.8789C31.2646 36.4424 40.4778 24.7254 50.7943 14.3576C52.0978 13.0476 55.8851 9.00298 54.7307 10.4554C53.3495 12.1931 52.1483 14.1379 50.9401 16.0066C44.7543 25.5739 38.1509 34.907 31.5742 44.1907C27.9506 49.3058 24.3105 54.4123 20.6155 59.4723C18.4551 62.4309 16.0429 65.4128 14.6259 68.8628C14.2583 69.758 14.4497 70.2298 15.2577 69.4419C17.006 67.737 18.5919 65.7748 20.2389 63.9662C24.3049 59.5014 28.491 55.1497 32.4003 50.535C38.1622 43.7334 43.4977 36.5636 49.2635 29.7651C52.3271 26.1528 55.3101 22.462 58.497 18.9647C59.1993 18.194 59.9017 17.4253 60.6352 16.6864C60.8693 16.4505 61.3399 15.6431 61.3399 15.9814C61.3399 16.9176 59.9435 18.4272 59.4811 19.141C57.2619 22.5664 55.164 26.0709 52.8597 29.4378C45.0836 40.7999 35.9684 51.1298 27.6621 62.0654C25.5881 64.7959 23.538 67.5701 21.9641 70.6503C21.4952 71.568 21.5323 72.0603 22.5351 71.4056C24.3941 70.1919 25.9297 68.498 27.4434 66.874C31.4432 62.5825 35.2706 58.1812 38.8394 53.4931C45.7824 44.3727 51.808 34.5652 58.6428 25.3594C58.7246 25.2491 59.9281 23.4159 59.2988 24.4027C58.1931 26.1366 57.28 28.0146 56.2372 29.7903C50.2461 39.9915 43.5209 49.7056 36.6161 59.2583C32.8705 64.4403 29.0081 69.5165 25.1594 74.6155C23.9996 76.152 22.8504 77.6968 21.7576 79.2855C21.3557 79.8697 20.677 81.4178 19.9595 81.7276C18.6009 82.3141 21.6246 79.2124 22.4865 77.9764C25.4085 73.7862 28.3543 69.6367 31.5985 65.7033C39.2123 56.4719 47.5663 47.9213 55.0952 38.6269C56.3606 37.0647 57.4676 35.2625 58.8615 33.8184C58.8889 33.7899 57.6032 36.6767 57.282 37.2926C52.5909 46.2896 47.4896 55.0165 42.4842 63.8277C40.5933 67.1564 38.6776 70.5034 37.2357 74.0742C36.7602 75.2518 35.3205 78.7318 36.0451 77.6995C39.9333 72.1603 44.1946 66.834 48.2673 61.436C54.6969 52.9142 60.6201 44.1544 66.2482 35.052C66.7571 34.2288 67.4633 32.7505 66.5641 34.5988C64.7796 38.2668 62.8951 41.8309 60.781 45.311C55.1595 54.5648 48.5749 63.1266 42.873 72.3119C42.701 72.589 39.8408 77.0485 40.2245 77.3218C40.6757 77.6433 44.3869 72.5214 44.5982 72.249C47.6047 68.3718 50.6575 64.5261 53.4186 60.4542C57.3981 54.5854 60.9148 48.3882 64.4501 42.227C65.6906 40.0652 66.8263 37.8234 68.0949 35.6814C68.9909 34.1685 66.9337 39.0255 66.2239 40.641C64.13 45.4068 62.0816 50.1864 59.6876 54.8022C56.7429 60.4798 53.3337 65.8131 50.0168 71.2545C49.3028 72.4257 48.6019 73.6106 47.9878 74.842C47.8004 75.218 47.1073 76.051 47.514 76.0127C47.7137 75.9939 48.51 74.7735 48.5346 74.7413C49.7448 73.1613 50.8896 71.5324 52.0093 69.8824C54.8355 65.7179 57.5324 61.4667 60.0278 57.0806C62.1928 53.2752 64.1737 49.2661 66.5155 45.5754C67.5708 43.9122 67.263 45.7733 67.0744 46.7083C66.2029 51.0278 64.4368 55.1697 62.5427 59.0947C60.602 63.1161 58.477 67.046 56.3951 70.9902C55.7335 72.2438 55.0845 73.5049 54.4148 74.7539C54.1087 75.3248 53.0454 76.8473 53.5887 76.5162C53.9468 76.2979 54.3544 75.3792 54.5363 75.0812C55.2342 73.9377 55.9305 72.7929 56.6017 71.6321C58.592 68.1901 60.3221 64.5481 62.5305 61.2472C62.6478 61.072 64.7443 58.1825 64.8389 58.8933C65.1721 61.3959 64.3464 64.2916 63.4539 66.5718C62.4144 69.2277 61.2231 71.8794 60.0521 74.477C59.7787 75.0834 59.4815 75.7248 59.1287 76.2896C59.0415 76.4293 59.5563 75.1934 59.639 75.0434C60.4224 73.6229 61.2018 72.203 61.996 70.7888C62.4323 70.0117 62.7418 68.9964 63.3324 68.3845"
         stroke="#D1AA77"
         stroke-width="2"
@@ -603,7 +616,7 @@
         position: absolute;
     }
 
-    .animate-bear-object {
+    .animate-bear-object-mobile {
         cursor: pointer;
     }
 </style>
